amidi-plug: Fix arithmetic overflow with files longer than 35 minutes (closes AUDPLUG-66)

+++ src/amidi-plug/i_midi.h
@@ -252,7 +252,7 @@ typedef struct
 
   gint playing_tick;
   gint avg_microsec_per_tick;
-  gint length;
+  gint64 length;
 
   gint skip_offset;
 }
